requirements.txt
# csv==0.0.1
# mysql-connector==2.2.9
# Flask==2.3.2
# getmac==0.9.3
# pip freeze
# python venv -m D:\GIT-HUB\venv
blinker==1.6.2
click==8.1.3       
colorama==0.4.6    
Flask==2.3.2       
getmac==0.9.3      
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
protobuf==3.20.3
Werkzeug==2.3.3